+++ Cargo.toml
@@ -5,7 +5,9 @@ members = [
     "data/schema_planner",
     "data_definition/operations",
     "data_definition/execution_plan",
-    "data_manipulation/query_plan",
+    "data_manipulation/annotated_queries",
+    "data_manipulation/operators",
+    "data_manipulation/untyped_tree",
     "deprecated/ast",
     "deprecated/binder_deprecated",
     "deprecated/catalog_deprecated/data_manager",
@@ -22,9 +24,7 @@ members = [
     "entities/types",
     "entities/definition",
     "meta_def",
-    "query_analysis/expr_operators",
     "query_analysis/query_analyzer",
-    "annotated_tree",
     "query_parsing/parser",
     "query_parsing/sql-ast",
     "query_validation/type_inference",
