--- annotated_tree/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "annotated_tree"
-version = "0.1.0"
-authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-data_definition_execution_plan = { path = "../data_definition/execution_plan" }
-expr_operators = { path = "../query_analysis/expr_operators" }
-types = { path = "../entities/types" }
-definition = { path = "../entities/definition" }
-sql_ast = { path = "../query_parsing/sql-ast" }
