+++ data_manipulation/untyped_tree/Cargo.toml
@@ -1,14 +1,12 @@
 [package]
-name = "validator"
+name = "data_manipulation_untyped_tree"
 version = "0.1.0"
 authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
 edition = "2018"
-publish = false
 
 [dependencies]
-data_manipulation_operators = { path = "../../data_manipulation/operators" }
-data_manipulation_untyped_tree = { path = "../../data_manipulation/untyped_tree" }
+data_manipulation_operators = { path = "../operators" }
 types = { path = "../../entities/types" }
 
 bigdecimal = { version = "0.2.0", features = ["string-only"] }
-log = "0.4.11"
+log = "0.4.13"
