+++ deprecated/schema_executor/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 publish = false
 
 [dependencies]
-annotated_tree = { path = "../../annotated_tree" }
 data_definition_operations = { path = "../../data_definition/operations" }
 data_definition_execution_plan = { path = "../../data_definition/execution_plan" }
