+++ query_analysis/query_analyzer/Cargo.toml
@@ -6,11 +6,11 @@ edition = "2018"
 publish = false
 
 [dependencies]
-annotated_tree = { path = "../../annotated_tree" }
 catalog = { path = "../../data/catalog" }
 data_definition_execution_plan = { path = "../../data_definition/execution_plan" }
-data_manipulation_query_plan = { path = "../../data_manipulation/query_plan" }
-expr_operators = { path = "../expr_operators" }
+data_manipulation_annotated_queries = { path = "../../data_manipulation/annotated_queries" }
+data_manipulation_operators = { path = "../../data_manipulation/operators" }
+data_manipulation_untyped_tree = { path = "../../data_manipulation/untyped_tree" }
 meta_def = { path = "../../meta_def" }
 definition = { path = "../../entities/definition" }
 types = { path = "../../entities/types" }
