+++ query_analysis/query_analyzer/src/tests/insertions/mod.rs
@@ -14,6 +14,7 @@
 
 use super::*;
 use data_definition_operations::{Kind, Record, Step, SystemObject, SystemOperation};
+use data_manipulation_untyped_tree::{ScalarValue, StaticEvaluationTree, StaticItem};
 
 #[cfg(test)]
 mod expressions;
