+++ query_analysis/query_analyzer/src/tests/updates/expressions.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use data_manipulation_untyped_tree::{DynamicEvaluationTree, DynamicItem, ScalarValue};
+
 use super::*;
 
 #[test]
@@ -175,6 +177,8 @@ fn update_table_with_parameters() {
 
 #[cfg(test)]
 mod multiple_values {
+    use data_manipulation_untyped_tree::{DynamicEvaluationTree, DynamicItem, ScalarValue};
+
     use super::*;
 
     fn update_value_as_expression_with_operation(
