+++ write_query/executor/Cargo.toml
@@ -6,8 +6,7 @@ edition = "2018"
 publish = false
 
 [dependencies]
-annotated_tree = { path = "../../annotated_tree" }
-data_manipulation_query_plan = { path = "../../data_manipulation/query_plan" }
+data_manipulation_annotated_queries = { path = "../../data_manipulation/annotated_queries" }
 catalog = { path = "../../data/catalog" }
 
 [dev-dependencies]
